continued with simulator

+++ src/main/java/ss/Start.java
@@ -21,37 +21,39 @@ public class Start {
 	public static void main(String[] args) {
 		Configuration config = Configuration.fromXML("configuracion.xml");
 		List<Project> projects = buildProjects(config);
-		long max = 0;
-		
-		//Get from projects maximum duration
-		for(Project p: projects) {
-			if(p.getDuration() > max) {
+		int max = 0;
+
+		// Get from projects maximum duration
+		for (Project p : projects) {
+			if (p.getDuration() > max) {
 				max = p.getDuration();
 			}
 		}
-		
+
 		ReasignationStrategy strategy = getStrategy(config.getStrategy());
-		
-		Simulator simulator = new SimulatorImpl(max, projects, config.getProgrammersQty(), strategy);
+
+		Simulator simulator = new SimulatorImpl(max, projects,
+				config.getProgrammersQty(), strategy);
 		simulator.start();
 	}
-	
+
 	private static ReasignationStrategy getStrategy(String strategy) {
-		//TODO: Instanciar el Distribution Manager
+		// TODO: Instanciar el Distribution Manager
 		return null;
 	}
 
 	private static List<Project> buildProjects(Configuration config) {
-		ss.gui.in.Project[] projects  = config.getProjects();
+		ss.gui.in.Project[] projects = config.getProjects();
 		List<Project> retList = Lists.newArrayList();
-		
+
 		for (int i = 0; i < projects.length; i++) {
 			retList.add(projects[i].buildProject());
 		}
-		
+
 		return retList;
-		
+
 	}
+
 	private static Properties loadProperties(String fileName) {
 		try {
 			Properties simulatorProperties = new Properties();
